% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confer_validation.R
\name{ACC_dt}
\alias{ACC_dt}
\title{Function for calculating anomaly correlation coefficients (ACCs) of ensemble mean forecasts stored in long data tables:}
\usage{
ACC_dt(
  fc_dt,
  fc_col,
  obs_dt = NULL,
  obs_col = ifelse(is.null(obs_dt), yes = "obs", no = fc_col),
  by_cols = intersect(c("month", "season", "lon", "lat", "system", "lead_time"),
    names(fc_dt)),
  along_cols = c("year"),
  ...
)
}
\arguments{
\item{fc_dt}{Data table containing the predictions.}

\item{fc_col}{column name of the prediction.}

\item{obs_dt}{Data table containing the observations. Pass NULL if the observation are contained in fc_dt.}

\item{obs_col}{column name of the observations (either in obs_dt, or in fc_dt if obs_dt = NULL).}

\item{by_cols}{column names of grouping variables, all of which need to be columns in fc_dt. A separate ACC is computed for each value of the grouping variables.
Default is to group by all instances of month, season, lon, lat, system and lead_time that are columns in fc_dt.}

\item{along_cols}{column name(s) for the variable(s) along which is averaged. Needs to contain 'year' per warning above.}

\item{...}{passed on to ACC_dt}
}
\description{
Can also handle point forecasts.
Warning: This metric always needs several years of data since the anomalies are calculated relative to multi-year means and standard deviations.
}
