% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_countries}
\alias{add_countries}
\title{adds country names to a data table containing lons and lats (on half or full degrees).
Only the 10 countries specified in the CONFER proposal are added (Sudan, South Sudan, Somalia, Eritrea, Ethiopia, Somalia, Kenya, Tansania, Uganda, Rwanda, Burundi)}
\usage{
add_countries(dt)
}
\arguments{
\item{dt}{the data table.}
}
\description{
adds country names to a data table containing lons and lats (on half or full degrees).
Only the 10 countries specified in the CONFER proposal are added (Sudan, South Sudan, Somalia, Eritrea, Ethiopia, Somalia, Kenya, Tansania, Uganda, Rwanda, Burundi)
}
