% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_country}
\alias{add_country}
\title{Same as add_country_names}
\usage{
add_country(dt, regions = EA_country_names())
}
\arguments{
\item{dt}{the data table.}

\item{regions}{Character vector of country names for which shapefiles are loaded.}
}
\description{
This is a synonyme for \code{\link{add_country_names}}.
Following a more intuitive naming convention, that is more in-line
with \code{add_climatology} and \code{add_tercile_cat}.
}
