% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_country_names}
\alias{add_country_names}
\title{Add country names to a data table with lon/lat coordinates}
\usage{
add_country_names(dt, regions = EA_country_names())
}
\arguments{
\item{dt}{the data table.}

\item{regions}{Character vector of country names for which shapefiles are loaded.
By default, countries in East Africa are loaded, see \code{\link{EA_country_names}}.
If you set regions = '.', the entire world is loaded, but this makes the function slower.}
}
\description{
Takes a data table with lon/lat coordinates and adds a column
'country' to it, containing the name of the country, the coordinate belongs to.

NOTE: We avoid dependencies on \pkg{sf} and \pkg{rnaturalearth}, which require gdal.
In particular, gdal is currently not available on ICPACs hpc, one of the key
environments where \pkg{SeaVal} is run in. We follow the 'older approach' relying
on the packages \pkg{maps}, \pkg{maptools}, and \pkg{sp} for getting shapefiles of
all countries.
}
