% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_tercile_cat}
\alias{add_tercile_cat}
\title{Add tercile category}
\usage{
add_tercile_cat(
  dt,
  datacol = c("obs", "prec", "precipitation", "pred"),
  by = setdiff(dimvars(dt), c("year", "member"))
)
}
\arguments{
\item{dt}{the data table.}

\item{datacol}{Name of the column where the data is stored. If multiple are provided, the first name matching a column in dt is used.}

\item{by}{names of columns to group by. Default is to group by everything except year and member, meaning that
the tercile categories are calculated seperately for each spatial coordinate and each month/season.}
}
\description{
Adds a column 'tercile_cat' with the tercile category to a data table with observations or predictions.
The tercile_cat is -1 for values falling in the lowest tercile (below normal), 0 for normal, and 1 for above normal.
terciles are calculated
}
