% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_tercile_probs}
\alias{add_tercile_probs}
\title{Adds columns 'below', 'normal' and 'above', containing predicted tercile probabilities, to a data table with ensemble forecasts.
The data table should either already have a column 'tercile_cat' (as would be added by \code{add_tercile_cat}),
or \code{add_tercile_cat} will be run first.}
\usage{
add_tercile_probs(dt, by = setdiff(dimvars(dt), "member"), ...)
}
\arguments{
\item{dt}{the data table.}

\item{by}{names of columns to group by}

\item{...}{passed on to \code{add_tercile_cat}.}
}
\description{
Adds columns 'below', 'normal' and 'above', containing predicted tercile probabilities, to a data table with ensemble forecasts.
The data table should either already have a column 'tercile_cat' (as would be added by \code{add_tercile_cat}),
or \code{add_tercile_cat} will be run first.
}
