% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_grids.R
\name{are_all_elements_within_eps}
\alias{are_all_elements_within_eps}
\title{Check if all elements of x are within tolerance eps of any element in y}
\usage{
are_all_elements_within_eps(x, y, eps)
}
\arguments{
\item{x}{A numeric vector, sorted in increasing order.}

\item{y}{A numeric vector, sorted in increasing order.}

\item{eps}{The tolerance within which we consider two values to be equal.}
}
\value{
A boolean value, TRUE if all x are in y within tolerance eps, FALSE otherwise.
}
\description{
Auxiliary function, used for checking whether spatial grids are regular, with allowing for rounding errors.
}
