% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confer_validation.R
\name{bootstrap_scores_dt}
\alias{bootstrap_scores_dt}
\title{Assess score uncertainty by bootstrapping}
\usage{
bootstrap_scores_dt(dt, r = 100, score_col, bycols = NULL, mc_cores = 1)
}
\arguments{
\item{dt}{Data table containing scores.}

\item{r}{how many bootstrap-resamples of the mean score should be considered?}

\item{score_col}{column name of the scores}

\item{by_cols}{should the mean score be aggregated? E.g. if you have a data table with scores from different systems,
putting 'system' here gives you a bootstrap estimate for the mean score of each system, which is usually what you want}
}
\description{
Assess score uncertainty by bootstrapping
}
