% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{by_cols_ens_fc_score}
\alias{by_cols_ens_fc_score}
\title{Auxiliary function returning the default column names to group by when calculating scores of ensemble forecasts.}
\usage{
by_cols_ens_fc_score(dt = NULL)
}
\arguments{
\item{dt}{optional. You can provide a data table, then the function returns the names of grouping variables in this data table.}
}
\description{
Auxiliary function returning the default column names to group by when calculating scores of ensemble forecasts.
}
