% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{by_cols_terc_fc_score}
\alias{by_cols_terc_fc_score}
\title{Auxiliary function}
\usage{
by_cols_terc_fc_score(dt = NULL)
}
\arguments{
\item{dt}{optional. You can provide a data table, then the function returns
the names of grouping variables in this data table.}
}
\value{
A vector of characters with the column names.
}
\description{
returns the default column names to group by when calculating scores for
tercile forecasts.
}
