% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{by_cols_terc_fc_score_sp}
\alias{by_cols_terc_fc_score_sp}
\title{Auxiliary function}
\usage{
by_cols_terc_fc_score_sp(dt = NULL)
}
\arguments{
\item{dt}{optional. You can provide a data table, then the function returns the names
of grouping variables in this data table.}
}
\value{
A vector of characters with the column names.
}
\description{
Gets column names to group by when calculating scores for tercile forecasts.
Some tercile forecasts, such as ROC score or SRC (slope of reliability curve)
require many data points and should therefore be pooled in space.
This auxiliary function returns the default column names to group by for these scores.
The suffix _sp stands for spatial pooling.
}
