% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{checks_ens_fc_score}
\alias{checks_ens_fc_score}
\title{Auxiliary function for scores of ensemble forecasts. Checks whether the data table contains columns with names that are not allowed,
or whether it is missing columns that are required.}
\usage{
checks_ens_fc_score()
}
\description{
Auxiliary function for scores of ensemble forecasts. Checks whether the data table contains columns with names that are not allowed,
or whether it is missing columns that are required.
}
