% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{checks_terc_fc_score}
\alias{checks_terc_fc_score}
\title{Auxiliary function for scores for tercile forecasts.}
\usage{
checks_terc_fc_score()
}
\description{
Checks whether the data table contains columns with names that are not allowed,
or whether it is missing columns that are required.
}
