% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{climatology_ens_forecast}
\alias{climatology_ens_forecast}
\title{Returns a leave-one-year-out climatology-based ensemble forecast}
\usage{
climatology_ens_forecast(obs_dt, by)
}
\arguments{
\item{obs_dt}{Data table containing observations, must contain a column 'year'.}

\item{by}{character vector containing the column names of the grouping variables, e.g. \code{c('month','lon','lat')}.}
}
\value{
Long data table with the typical ensemble-forecast looks, i.e. containing a column 'member'.
}
\description{
For a given year, the ensemble forecast simply consists of the observations in all other years.
In particular, the ensemble size is (number of years in the data table) - 1.
This is essentially an auxiliary function for computing skill scores.
}
