% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{climatology_threshold_exceedence}
\alias{climatology_threshold_exceedence}
\title{Get climatological prediction for exceedence probabilities.}
\usage{
climatology_threshold_exceedence(
  obs_dt,
  o = "prec",
  by = setdiff(coords(obs_dt), "year"),
  thresholds = c(200, 300, 350, 400)
)
}
\arguments{
\item{obs_dt}{Data table containing observations.}

\item{thresholds}{vector of thresholds for which the exceedence probabilities should be derived.}

\item{obs_col}{column name of the observation. Mostly observed precipitation in mm.}

\item{by_cols}{By which columns should be grouped?}
}
\description{
The climatological prediction for exceedence probabilities is the fraction of observed years where the observation exceeded the threshold.
It's calculated from leave-one-year-out climatology.
}
