% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine}
\alias{combine}
\title{Combine two data tables}
\usage{
combine(dt1, dt2, ...)
}
\arguments{
\item{dt1}{first data table}

\item{dt2}{second data table}

\item{...}{passed on to data.table::merge}
}
\description{
Convenience wrapper for data.tables merge. Merges by dimvars and returns user-friendly warnings.
}
