% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confer_validation.R
\name{composite_analysis}
\alias{composite_analysis}
\title{composite analysis for teleconnections}
\usage{
composite_analysis(
  var_dt,
  TC_dt,
  TC_name = names(TC_dt)[3],
  by_cols = c("month", "lon", "lat"),
  average_along_cols = "year",
  var_name = setdiff(names(var_dt), c(average_along_cols, by_cols))[1]
)
}
\arguments{
\item{var_dt}{Data table containing the weather variables for which you want to conduct the composite analysis, e.g. monthly mean precip for a range of years, months, and locations.}

\item{TC_dt}{Data table containing the values of the teleconnection index (such as IOD), typically indexed by year and month.}

\item{TC_name}{Character string containing the column name of TC_dt where the values of the teleconnection index are stored. Default is third column.}

\item{by_cols}{Character vector containing the column names of the grouping variables by which the composite analysis is supposed to be conducted. Default is \code{c('month','lon','lat')}.}

\item{var_name}{Column name of the weather variable in var_dt. Default is first column not contained in \code{by_cols} and not named 'year'.}
}
\value{
A data table containing the composites x_plus and x_minus for each value of by_cols
}
\description{
composite analysis for teleconnections
}
\author{
Claudio
}
