% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{coords}
\alias{coords}
\title{Return the default column names that are considered to be coordinates rather than values:}
\usage{
coords(dt = NULL)
}
\description{
Return the default column names that are considered to be coordinates rather than values:
}
