% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_plot.R
\name{countryplot_dt}
\alias{countryplot_dt}
\title{plotting spatial data for one country}
\usage{
countryplot_dt(dt, data_col = colnames(dt)[3], country, tol = 0, ...)
}
\arguments{
\item{dt}{Data table containing the data for plotting.}

\item{data_col}{The name of the column in dt containing the data for plotting.}

\item{country}{Name of the country to plot. Possible values are: "Burundi", "Djibouti", "Eritrea", "Ethiopia", "Kenya", "Rwanda", "Somalia", "South Sudan", "Sudan", "Tanzania", "Uganda"}

\item{tol}{length by which the plot window is widened (in lon/lat, so tol = 0.5 includes roughly one pixel beyond the country borders, when your data is on a half degree grid).}

\item{...}{passed on to ggplot_dt}
}
\value{
a ggplot object.
}
\description{
This function wraps \code{ggplot_dt}.
}
\author{
Claudio Heinrich
}
