% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{CPA_dt}
\alias{CPA_dt}
\title{Function changed name, see CPA}
\usage{
CPA_dt(
  dt,
  f,
  o = "obs",
  by = by_cols_ens_fc_score(dt),
  pool = "year",
  mem = "member",
  dim.check = TRUE
)
}
\description{
Function changed name, see CPA
}
