% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R
\name{create_diagram_by_level}
\alias{create_diagram_by_level}
\title{Auxiliary function to simplify grouping for diagrams}
\usage{
create_diagram_by_level(FUN, by, dt, ...)
}
\arguments{
\item{FUN}{The name of the function creating the diagram}

\item{by}{Column names in dt to group by}

\item{dt}{data table (cannot be part of ..., because a sub-data-table is passed to FUN)}

\item{...}{arguments passed to FUN}
}
\description{
Only works for functions that return a single plot if by == NULL.
This is not the case for some functions plotting results for all three categories, e.g. reliability diagrams or ROC curves.
}
