% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{crps_aux}
\alias{crps_aux}
\title{Auxiliary function for calculating crps.}
\usage{
crps_aux(y, dat)
}
\arguments{
\item{y}{vector of length m with m identical entries, the observation}

\item{dat}{vector of length m containing the m ensemble forecasts}
}
\description{
Mostly copy-paste from \code{scoringRules:::crps_edf}. Adjusted to the data table format, where the observation is a vector of the same length as the ensemble forecast,
but is just repeated (which is why only \code{y[1]}) is used.
}
