% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{CRPSS_ens_fc}
\alias{CRPSS_ens_fc}
\title{Function got renamed, please see CRPSS}
\usage{
CRPSS_ens_fc(
  dt,
  f,
  o = "obs",
  by = by_cols_ens_fc_score(),
  pool = c("year"),
  ...
)
}
\description{
Function got renamed, please see CRPSS
}
