% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{cv_to_dt}
\alias{cv_to_dt}
\title{function for loading cross-validation data (predictions and observations).}
\usage{
cv_to_dt(data_dir)
}
\arguments{
\item{data_dir}{path to where the cross-validation is stored.}
}
\value{
A data table.
}
\description{
function for loading cross-validation data (predictions and observations).
}
\examples{
{
fn = '/nr/user/claudio/bigdisk/SFE/ERA_monthly_nc/total_precipitation_era_1979_1.nc'
ncdf_to_dt(fn, subset_list = list('latitude' = -50:50)) # will read out everything between lat -50 and lat 50 (not only the latitudes matching -50:50)
}

}
\author{
Claudio
}
