% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{data_dir}
\alias{data_dir}
\title{Auxiliary function to access the directory used to load and save data. The first time this is called it asks the user to configure it.}
\usage{
data_dir(set_dir = F)
}
\arguments{
\item{set_dir}{logical. Set this to TRUE if you have to redefine your data directory.}
}
\description{
Auxiliary function to access the directory used to load and save data. The first time this is called it asks the user to configure it.
}
