% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirps.R
\name{delete_redundant_files}
\alias{delete_redundant_files}
\title{Auxiliary function cleaning out the directories}
\usage{
delete_redundant_files(dir)
}
\description{
Auxiliary function cleaning out the directories
}
