% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{dimvars}
\alias{dimvars}
\title{Return the default column names that are considered to be coordinates rather than values:}
\usage{
dimvars(dt = NULL)
}
\arguments{
\item{dt}{Optional data table. If a data table is provided only the dimvars of the data table are returned.}
}
\description{
Return the default column names that are considered to be coordinates rather than values:
}
