% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirps.R
\name{download_chirps_monthly_high}
\alias{download_chirps_monthly_high}
\title{Auxiliary function called by download_chirps_monthly}
\usage{
download_chirps_monthly_high(
  update,
  version,
  years,
  months,
  extent,
  timeout_limit,
  save_dir = file.path(chirps_dir(), version)
)
}
\description{
Auxiliary function called by download_chirps_monthly
}
