% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{EA_country_names}
\alias{EA_country_names}
\title{Get names of countries in east Africa}
\usage{
EA_country_names()
}
\description{
This is an auxiliary function used in \link{add_country_names}, so only these names are recognized
by default.
}
\details{
@examples
EA_country_names()
}
