% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{fc_cols}
\alias{fc_cols}
\title{Forecast column names}
\usage{
fc_cols(dt = NULL)
}
\arguments{
\item{dt}{optional data table. If provided, the function guesses which column contains the forecast values. Else it returns all recognized forecast column names.}
}
\value{
Character vector with column names.
}
\description{
returns the columns names that are recognized as (ensemble-) forecast values
}
\examples{
fc_cols()

}
