% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot_dt}
\alias{ggplot_dt}
\title{plotting function for spatial data}
\usage{
ggplot_dt(
  dt,
  data_col = NULL,
  mn = NULL,
  discrete_cs = FALSE,
  rr = NULL,
  low = "blue",
  mid = "white",
  high = "red",
  name = data_col,
  midpoint = NULL,
  na.value = "gray50",
  oob = scales::squish,
  guide = guide_colorbar(barwidth = 0.5, barheight = 10),
  ...,
  binwidth = NULL,
  bin_midpoint = midpoint
)
}
\arguments{
\item{dt}{Data table containing the data for plotting.}

\item{data_col}{The name of the column in dt containing the data for plotting. If nothing is provided (the default), the first column that is not a dimension variable or 'member' is selected.}

\item{mn}{optional plot title}

\item{discrete_cs}{Logical. Should the color scale be discretized?}

\item{rr, low, mid, high, name, midpoint, na.value, oob, guide, ...}{Arguments for the color scale, passed to scale_fill_gradient2 or scale_fill_steps2 (depending on whether discrete_cs == TRUE).
rr replaces limits (specifying the range of the color scale) for consistency with the older plotting functions from the PostProcessing package.
The default midpoint is the center of the data range (or the center of rr, if provided), not 0 (as is the default for scale_fill_gradient2).
Specifying the midpoint can often be a convenient way to force a color scale with only two colors (for example, by setting it
to the minimum or maximum of your data). \code{na.value} specifies the color of missing values. \code{oob} specifies the treatment of out-of-boundary values, i.e. values beyond the
limits.
The ... argument can in particular be used to customize the legend/colorbar using the function \code{guide_colorbar},
see https://ggplot2.tidyverse.org/reference/guide_colourbar.html. Moreover, when \code{discrete_cs == TRUE} you can pass the arguments \code{n.breaks,breaks} to customize the scale.
If you use n.breaks you might also want to set nice.breaks = FALSE, see ?scale_fill_steps2.}

\item{binwidth, bin_midpoint}{only used when \code{discrete_cs == TRUE}. Normally, the breaks for the discrete colorscale are specified by n.breaks (which is not reliable,
since they're adjusted to be 'nice'), or by specifying the breaks explicitly (which is often tedious). This gives you a third option, namely specifying how far the breaks
should be apart, and specifying the centerpoint for one of the bins (default is midpoint, or the center of rr if midpoint is not provided). For example, if your color scale shows
percentages and you'd like 4 categories, ranging from white to red, this is easiest achieved by \code{binwidth = 25, midpoint = 12.5}.}
}
\value{
a ggplot object.
}
\description{
Plots spatial data from a data.table. The data table needs to contain columns named 'lon' and 'lat'. The grid needs to be regular.
If spatial data is contained for several levels (e.g. mutliple times or multiple ensemble members), only the first is plotted.
}
\author{
Claudio Heinrich
}
