% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot_dt_gha_map}
\alias{ggplot_dt_gha_map}
\title{Plotting function with different map for Greater Horn of Africa}
\usage{
ggplot_dt_gha_map(..., expand.x = c(-0.5, 0.5), expand.y = c(-0.5, 2))
}
\arguments{
\item{..., expand.x, expand.y}{passed to \code{\link{ggplot_dt}}}
}
\description{
This essentially wraps \code{\link{ggplot_dt}}, but uses a different map for borders.
The map is part of the package and is the one currently used during GHACOFs at ICPAC.
}
\examples{
dt = chirps_monthly[lon \%between\% c(30,40) & lat < 0 & month == 11 & year == 2020]
pp = ggplot_dt_gha_map(dt)
if(interactive()) plot(pp)

}
