% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirps.R
\name{GHA_extent}
\alias{GHA_extent}
\title{GHA-bounding-box}
\usage{
GHA_extent()
}
\value{
A numeric vectorof length 4.
}
\description{
Returns a lon/lat bounding box for the greater horn of Africa region. Format is c(xmin,xmax,ymin,ymax), as for raster::extent
}
\examples{
GHA_extent()
}
