% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_grids.R
\name{grid_info}
\alias{grid_info}
\title{Retrieve spatial grid information from a data table}
\usage{
grid_info(dt)
}
\arguments{
\item{dt}{A data table}
}
\value{
This function does not return a value; instead, it prints a message to the console
with the grid information.
}
\description{
This function prints out spatial grid information from a data table. If the grid-attribute does not exist
\code{\link{set_spatial_grid}} is called first.
}
\examples{

 dt = data.table(lon = runif(10), lat = runif(10))
 grid_info(dt)
}
