% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{lead_time_conversion}
\alias{lead_time_conversion}
\title{Adjust Lead Time for Seasonal Forecasts}
\usage{
lead_time_conversion(lead_time, month)
}
\arguments{
\item{lead_time}{A numeric vector of lead time values.}

\item{month}{A numeric vector of month values.}
}
\value{
A numeric vector of lead time values adjusted for seasonal forecasts.
}
\description{
This function adjusts the lead time values from monthly data for the seasonal format.
}
\details{
The function adjusts the lead time values by subtracting the difference between the month value and the corresponding starting month of the season.
The adjustment is performed for the months 3-5 (March to May), 6-9 (June to September), and 10-12 (October to December).
For example, if the month is March (3), the lead time is adjusted by subtracting (3 - 3) = 0.
This adjustment aligns the lead time values with the seasonal forecasts.
}
