% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirps.R
\name{load_chirps}
\alias{load_chirps}
\title{Function for reading in (previously downloaded) CHIRPS (monthly) data.}
\usage{
load_chirps(years = NULL, months = NULL, us = T)
}
\arguments{
\item{years, months}{Optional subset of years and months you want to load}

\item{us}{logical. If TRUE, the upscaled version is loaded.}
}
\value{
the derived data table
}
\description{
The resulting data table contains precip in unit mm/day.
}
