% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{lt_cols}
\alias{lt_cols}
\title{Data table column names that are recognized as leadtime}
\usage{
lt_cols()
}
\description{
Data table column names that are recognized as leadtime
}
\examples{
lt_cols()
}
