% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICPAC_temp.R
\name{mask_dt}
\alias{mask_dt}
\title{Function for applying a mask derived by \link{\code{get_mask}}}
\usage{
mask_dt(dt, mask)
}
\arguments{
\item{dt}{Data table you want to apply the masking to}

\item{mask}{A masking data table generated by \link{\code{get_mask}}}
}
\description{
Function for applying a mask derived by \link{\code{get_mask}}
}
