% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{MBSS_dt}
\alias{MBSS_dt}
\title{Function got renamed, please see MBSS}
\usage{
MBSS_dt(
  dt,
  f = c("below", "normal", "above"),
  o = "tercile_cat",
  by = by_cols_terc_fc_score(),
  pool = "year",
  dim.check = TRUE
)
}
\description{
Function got renamed, please see MBSS
}
