% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{modify_dt_map_plotting}
\alias{modify_dt_map_plotting}
\title{Auxiliary function for checking dimensions for map-plotting}
\usage{
modify_dt_map_plotting(dt, data_col)
}
\arguments{
\item{dt}{Data table containing the data for plotting}

\item{data_col}{Name of column containing the data for plotting}
}
\description{
Auxiliary function for checking dimensions for map-plotting
}
