% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf_to_dt.R
\name{ncdf_to_dt}
\alias{ncdf_to_dt}
\title{function for converting netcdfs to long data tables.}
\usage{
ncdf_to_dt(nc, subset_list = NULL, printunits = TRUE)
}
\arguments{
\item{nc}{either character string with the name of the .nc file (including path), or an object of type ncdf4}

\item{subset_list}{named list for subsetting. The names must match dimension names of the nc file. The range of the vector contained on each page defines which subset of the data is read out.}

\item{printunits}{Logical: Should the units be printed}
}
\value{
long data table.
}
\description{
This function is deprecated, use netcdf_to_dt instead. It crashes when the netcdf has 'empty' dimension variables that are not used by any variable.
This is the case for some netcdfs provided by ICPAC and CORDEX.
}
\author{
Claudio
}
