% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf_to_dt.R
\name{netcdf_to_dt}
\alias{netcdf_to_dt}
\title{function for converting netcdfs to long data tables.}
\usage{
netcdf_to_dt(nc, vars = NULL, verbose = 2, trymerge = TRUE, subset_list = NULL)
}
\arguments{
\item{nc}{either character string with the name of the .nc file (including path), or an object of type ncdf4.}

\item{vars}{Which variables should be read from the netcdf? Either a character vector of variable names, or a
integer vector of variable indices, or NULL in which case all variables are read.}

\item{verbose}{Either 0, 1 or 2. How much information should be printed?
The default (2) is to print the entire netcdf information, 1 just prints the units of all variables, 0 (or any other input)
prints nothing.}

\item{trymerge}{logical. If TRUE a single data table containing all variables is returned, else a list of data
tables, one for each variable. The latter is more memory efficient if you have multiple variables that depend
on different dimensions.}

\item{subset_list}{A named list. The names of the pages must correspond to the names of dimension variables in the netcdf, and each page contains a (two-element-)range vector.
For example, subsetting a global dataset to just East Africa could look like this: subset_list = list(latitude = c(-15,25),longitude = c(20,55)).
Non-rectangular subsetting during reading a netcdf seems to be difficult, see ncvar_get. Every dimension variable not named in subset_list is read entirely.}
}
\value{
A data table if \code{trymerge == TRUE} or else a list of data tables.
}
\description{
The function converts netcdfs into data.tables.
Be aware that the data table can be much larger in memory, especially if you have many dimension variables.
This function is more robust than the older version \code{ncdf_to_dt}. In particular it does not get confused
if 'empty' dimension variables are present (some files from ICPAC and CORDEX).
}
\examples{
{
fn = '/nr/project/stat/CONFER/Data/validation/example_data/202102/CorrelationSkillRain_Feb-Apr_Feb2021.nc'
test = netcdf_to_dt(nc)
}

}
\author{
Claudio
}
