% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{obs_coords}
\alias{obs_coords}
\title{Auxiliary function returning all column names indicating an observation-coordinate, i.e. a coordinate for which only one observation may exist.}
\usage{
obs_coords(dt = NULL)
}
\arguments{
\item{dt}{optional. You can provide a data table, then the function returns the names of spatial coordinate columns in this data table.}
}
\description{
Auxiliary function returning all column names indicating an observation-coordinate, i.e. a coordinate for which only one observation may exist.
}
