% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confer_validation.R
\name{PCC_dt}
\alias{PCC_dt}
\title{Function for calculating Pearson correlation coefficients (PCCs) of ensemble mean forecasts stored in long data tables:}
\usage{
PCC_dt(
  fc_obs_dt,
  fc_col,
  obs_col = "obs",
  by_cols = intersect(c("month", "season", "lon", "lat", "system", "lead_time"),
    names(fc_obs_dt)),
  along_cols = c("year")
)
}
\arguments{
\item{fc_obs_dt}{Data table containing the predictions.}

\item{fc_col}{column name of the prediction.}

\item{obs_col}{column name of the observations.}

\item{by_cols}{column names of grouping variables, all of which need to be columns in fc_dt. A separate PCC is computed for each value of the grouping variables.
Default is to group by all instances of month, season, lon, lat, system and lead_time that are columns in fc_obs_dt.}

\item{along_cols}{column name(s) for the variable(s) along which is averaged. Needs to contain 'year' per warning above.}

\item{...}{passed on to PCC_dt}
}
\description{
Can also handle point forecasts.
Warning: This metric always needs several years of data since the means and standard deviations are calculated across time.
}
