% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{RES}
\alias{RES}
\title{Compute Resolution scores}
\usage{
RES(
  dt,
  bins = c(0.3, 0.35001),
  f = c("below", "normal", "above"),
  o = "tercile_cat",
  by = by_cols_terc_fc_score(),
  pool = "year",
  dim.check = TRUE
)
}
\arguments{
\item{dt}{Data table containing the predictions.}

\item{bins}{probability bins, defaults to c("<30", "30-35",">35")}
}
\description{
Computes both the resolution component of the Brier score or resolution component of the Ignorance score.
Mason claims to prefer the ignorance score version, but this has a very high chance of being NA (much higher
than for the full ignorance score itself, I think we should drop it for that reason). Mason writes that the
scores are unstable for single locations and that one should pool over many locations.
Requires the specification of probability bins. One score for each category (below, normal, above) and
also the sum of the scores.
}
\details{
Values close to 0 means low resolution. Higher values mean higher resolution.
}
