% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restrict_to_confer_region}
\alias{restrict_to_confer_region}
\title{Restrict data to the region typically considered in CONFER}
\usage{
restrict_to_confer_region(dt, ...)
}
\arguments{
\item{dt}{the data table.}

\item{...}{passed on to restrict_to_country}
}
\description{
Wraps \code{restrict_to_country}, and restricts to the GHA-region usually considered in the Horizon2020 project CONFER, consisting
of the following countries: 'Burundi','Djibouti', 'Eritrea','Ethiopia','Kenya','Rwanda','Somalia','Somaliland','South Sudan','Sudan','Tanzania','Uganda'
}
