% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restrict_to_country}
\alias{restrict_to_country}
\title{restricts data to a specified country}
\usage{
restrict_to_country(dt, ct, rectangle = FALSE, tol = 0.5)
}
\arguments{
\item{dt}{the data table.}

\item{ct}{name of the country, or vector containing multiple country names}

\item{rectangle}{logical. If FALSE (default), the data is restricted to the gridcells for which the centerpoint lies within the selected country (e.g. for computing mean scores for a country).
If TRUE, the data is kept for a rectangle containing the entire country, therefore also containing gridpoints outside the country. This is the preferred option for plotting data
for a specific country.}

\item{tol}{Only used when \code{rectangle == TRUE}. A tolerance value for widening the plotting window, making things look a bit nicer.}
}
\description{
Restricts a dataset to one or more countries, specified by their names. If you have lon/lat data and don't know which countries these coordinates belong to, see \code{add_country_names}.
}
