% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restrict_to_GHA}
\alias{restrict_to_GHA}
\title{restricts data to the Greater Horn of Africa}
\usage{
restrict_to_GHA(dt, ...)
}
\arguments{
\item{dt}{the data table.}

\item{...}{passed on to \code{\link{restrict_to_country}}}
}
\value{
the data table, restricted to the selected country
}
\description{
Wraps \code{\link{restrict_to_country}}, and restricts to the GHA-region usually considered in CONFER, see \code{\link{EA_country_names}}.
}
\examples{
ex_dt = chirps_monthly[lat < 0 & month == 11 & year == 2020]
dt = restrict_to_GHA(ex_dt)

}
