% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{run_dimension_check_terc_forecast}
\alias{run_dimension_check_terc_forecast}
\title{Auxiliary Function called inside functions that calculate scores
for ensemble forecasts. Checks whether the provided data table has the right format.}
\usage{
run_dimension_check_terc_forecast()
}
\description{
Auxiliary Function called inside functions that calculate scores
for ensemble forecasts. Checks whether the provided data table has the right format.
}
