% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{season}
\alias{season}
\title{Getting seasons from season}
\usage{
season(mm)
}
\arguments{
\item{mm}{vector of months}
}
\description{
takes a vector of months and returns a same-length vector of seasons
}
