% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{season_months}
\alias{season_months}
\title{Getting months from season}
\usage{
season_months(ss)
}
\arguments{
\item{ss}{vector containing one or multiple season names ('MAM', 'JJAS', 'OND')}
}
\description{
takes a vector of up to three seasons and returns the months in them
}
