% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{season_strings_to_int}
\alias{season_strings_to_int}
\title{Auxiliary function for decoding season-strings}
\usage{
season_strings_to_int(seasons = c("MAM", "JJAS", "OND"))
}
\arguments{
\item{seasons}{A character vector of season strings, see \code{\link[=convert_monthly_to_seasonal]{convert_monthly_to_seasonal()}} for details}
}
\description{
Auxiliary function for decoding season-strings
}
