% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonal}
\alias{seasonal}
\title{Convert Data to Seasonal Format}
\usage{
seasonal(dt, vars = NULL, by = NULL, FUN = sum, subset = TRUE)
}
\arguments{
\item{dt}{The input data table.}

\item{vars}{A character vector specifying the columns to be converted (optional).}

\item{by}{A character vector specifying the grouping variables (optional).}

\item{FUN}{The function to be used for aggregation (default is \code{sum}).}

\item{subset}{logical. If TRUE (default), then we expect three entries per level of by in MAM and OND, and 4 levels in JJAS.
The data is subset to all levels of by satisfying this condition. This is important if your data may be incomplete for some
levels. For example, if you have monthly precip data and it ends in October 2023, then the OND-2023 precip entries would just be the October values,
because the function simply sums over all OND 2023 values it finds. This is usually not what you want, you'd prefer OND 2023 to be suppressed, which
is the functions behavior when subset = TRUE.}
}
\value{
A data table in seasonal format with aggregated values.
}
\description{
This function converts data to seasonal format by aggregating monthly values.
}
\details{
The function checks if the input data table is already in seasonal format.
If not, it requires the data table to contain a column named "month" for conversion.
The function then aggregates the data by season based on the month column and the specified grouping variables.
The specified columns are reduced to those that are numeric-like (numeric or logical).
The function returns a new data table with the aggregated values.
}
