% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{space_dimvars}
\alias{space_dimvars}
\title{Auxiliary function returning all column names indicating a spatial coordinate.}
\usage{
space_dimvars(dt = NULL)
}
\arguments{
\item{dt}{optional. You can provide a data table, then the function returns the names of spatial coordinate columns in this data table.}
}
\description{
Auxiliary function returning all column names indicating a spatial coordinate.
}
