% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{tc_cols}
\alias{tc_cols}
\title{which column names are interpreted as obeserved tercile categories}
\usage{
tc_cols(dt = NULL)
}
\arguments{
\item{dt}{optional data table. If provided, the function guesses which column contains the observations. Else it returns all recognized column names.}
}
\description{
which column names are interpreted as obeserved tercile categories
}
