% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICPAC_temp.R
\name{tercile_plot}
\alias{tercile_plot}
\title{Function for plotting terciles}
\usage{
tercile_plot(
  dt,
  data_col = tc_cols(dt),
  mn = NULL,
  low = "orange",
  mid = "cyan",
  high = "green1",
  name = "",
  labels = c("Wetter", "Average", "Drier"),
  na.value = "white",
  extent = NULL,
  expand.x = c(-0.5, 0.5),
  expand.y = c(-0.5, 2)
)
}
\arguments{
\item{dt}{data table}

\item{data_col}{Name of the column containing the observed tercile category}

\item{mn}{optional title for the plot.}

\item{low, mid, high}{colors for the three categories}

\item{name}{optional title for the colorscale}

\item{labels}{How to label the three categories}

\item{na.value}{How to color missing values}
}
\description{
Function for plotting terciles
}
