% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{time_dimvars}
\alias{time_dimvars}
\title{Auxiliary function}
\usage{
time_dimvars(dt = NULL)
}
\arguments{
\item{dt}{optional. You can provide a data table, then the function returns the names of temporal coordinate columns in this data table.}
}
\value{
Character vector with column names.
}
\description{
returns all column names indicating a temporal coordinate.
}
\examples{
time_dimvars()

}
