% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upscale_to_half_degrees}
\alias{upscale_to_half_degrees}
\title{This function is deprecated! Please use ForecastTools::upscale_regular_lon_lat instead.}
\usage{
upscale_to_half_degrees(
  dt,
  uscol,
  bycols = intersect(c("month", "year"), colnames(dt))
)
}
\description{
This function is deprecated! Please use ForecastTools::upscale_regular_lon_lat instead.
}
