% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_country}
\alias{add_country}
\title{Same as add_country_names}
\usage{
add_country(dt, regions = EA_country_names())
}
\description{
This is a synonyme for add_country_names.
Following a more intuitive naming convention, that is more in-line
with add_climatology and add_tercile_cat
}
