% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_country_names}
\alias{add_country_names}
\title{Takes a data table with lon/lat coordinates and returns the same data table with a new column 'country' containing the countryname
of each coordinate. Usually you'd do this using sf, but that introduces dependency to gdal which we want to avoid.}
\usage{
add_country_names(dt, load_continent = "Africa")
}
\arguments{
\item{dt}{the data table.}

\item{load_continent}{which continent to load. Loading only one continent makes the function faster.
Put NULL if you have data across several continents}
}
\description{
Takes a data table with lon/lat coordinates and returns the same data table with a new column 'country' containing the countryname
of each coordinate. Usually you'd do this using sf, but that introduces dependency to gdal which we want to avoid.
}
