% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_country_names}
\alias{add_country_names}
\title{adds a column with the countryname to a data table
Only the following 10 countries are added: Sudan, South Sudan, Somalia, Eritrea, Ethiopia, Somalia, Kenya, Tansania, Uganda, Rwanda, Burundi}
\usage{
add_country_names(dt)
}
\arguments{
\item{dt}{the data table.}
}
\description{
adds a column with the countryname to a data table
Only the following 10 countries are added: Sudan, South Sudan, Somalia, Eritrea, Ethiopia, Somalia, Kenya, Tansania, Uganda, Rwanda, Burundi
}
