% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_country_names}
\alias{add_country_names}
\title{Add country names to a data table}
\usage{
add_country_names(dt, load_continent = "Africa")
}
\arguments{
\item{dt}{the data table.}

\item{load_continent}{which continent to load. Loading only one continent makes the function faster. Default is Africa.
Put NULL if you have data across several continents.}
}
\description{
Takes a data table with lon/lat coordinates and returns the same data table with a new column 'country' containing the country name
of each coordinate.
}
