% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_tercile_cat}
\alias{add_tercile_cat}
\title{adds a column with the observed tercile category to a data table}
\usage{
add_tercile_cat(
  dt,
  datacol = NULL,
  years = NULL,
  by = setdiff(dimvars(dt), c("year", "member"))
)
}
\arguments{
\item{dt}{the data table.}

\item{datacol}{Name of the column where the data is stored. If NULL, the function guesses.}

\item{years}{Optional, if provided only these years are used for establishing climatology terciles.}

\item{by}{names of columns to group by.}
}
\description{
adds a column with the observed tercile category to a data table
}
