% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_tercile_cat}
\alias{add_tercile_cat}
\title{adds a column with the tercile category to a data table}
\usage{
add_tercile_cat(
  dt,
  datacol = "prec",
  by = intersect(c("month", "lon", "lat"), names(dt))
)
}
\arguments{
\item{dt}{the data table.}

\item{datacol}{Name of the column where the data is stored}

\item{by}{names of columns to group by}
}
\description{
adds a column with the tercile category to a data table
}
