% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confer_validation.R
\name{BSS_ex_dt}
\alias{BSS_ex_dt}
\title{Calculate exceedence Brier skill score}
\usage{
BSS_ex_dt(
  fc_dt,
  fc_col,
  clim_col = "clim",
  threshold_col,
  obs_col = "obs",
  by_cols = intersect(c("month", "season", "lon", "lat", "system", "lead_time"),
    names(fc_dt))
)
}
\arguments{
\item{fc_dt}{Data table containing the predictions.}

\item{fc_col}{column name of the prediction. Contains predicted probabilities of exceedence}

\item{clim_col}{column name containing the out-of-sample climatology forecast for the exceedence probability.
Usually generated by \code{climatology_threshold_exceedence}.}

\item{threshold_col}{which column contains the exceedence threshold?}

\item{obs_col}{column name of the observations. Can either be logical (exceedence or not) or real valued, containing precipitation (or the variable for which exceedence should be checked).}

\item{by_cols}{column names of grouping variables, all of which need to be columns in fc_dt.
Default is to group by all instances of month, season, lon, lat, system and lead_time that are columns in fc_dt.}
}
\description{
Calculate exceedence Brier skill score
}
