% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirps.R
\name{chirps_dir}
\alias{chirps_dir}
\title{Auxiliary function to access/set the directory for loading and saving CHIRPS data.}
\usage{
chirps_dir(dir = paste0(data_dir(), "CHIRPS/"))
}
\arguments{
\item{dir}{The directory}
}
\description{
Auxiliary function to access/set the directory for loading and saving CHIRPS data.
}
