% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirps.R
\name{chirps_dir}
\alias{chirps_dir}
\title{CHIRPS directory}
\usage{
chirps_dir(dir = file.path(data_dir(), "CHIRPS"))
}
\arguments{
\item{dir}{The directory}
}
\value{
The directory path.
}
\description{
Auxiliary function to access/set the directory for loading and saving CHIRPS data.
}
\examples{
if(interactive()){chirps_dir()}

}
