% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chirps_monthly}
\alias{chirps_monthly}
\title{Monthly mean precipitation}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 209040 rows and 6 columns.
}
\source{
\url{http://iridl.ldeo.columbia.edu/SOURCES/.UCSB/.CHIRPS/.v2p0/.monthly/.global/.precipitation/}
}
\usage{
data(chirps_monthly)
}
\description{
This dataset contains observed monthly mean precipitation for the greater horn of Africa,
for November - December 1991-2020. The unit of precipitation is mm/day. It also contains the tercile category, where -1
means below normal rainfall (lowest tercile for this location and month), 0
is normal and 1 is above normal.The data source is CHIRPS-blended, upscaled
to a half-degree grid.
}
\keyword{datasets}
