% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine}
\alias{combine}
\title{Combine two data tables}
\usage{
combine(dt1, dt2, ...)
}
\arguments{
\item{dt1}{first data table}

\item{dt2}{second data table}

\item{...}{passed on to data.table::merge}
}
\description{
Function for combining two data tables, e.g. with predictions and observations.
This is essentially a user-friendly wrapper for data.tables merge that guesses the columns to merge by (the dimension variables
contained in both data tables.
}
