% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confer_validation.R
\name{CRPS_ens_fc}
\alias{CRPS_ens_fc}
\title{Taking CRPSs of ensemble forecasts stored in long data tables:}
\usage{
CRPS_ens_fc(
  fc_dt,
  fc_cn,
  obs_dt = NULL,
  obs_cn = ifelse(is.null(obs_dt), yes = "obs", no = fc_cn),
  by_cns = intersect(c("month", "season", "lon", "lat", "system", "lead_time"),
    names(fc_dt)),
  along_cns = c("year"),
  check_dimensions = T,
  member_cn = "member"
)
}
\arguments{
\item{fc_dt}{Data table containing the predictions.}

\item{fc_cn}{column name of the prediction.}

\item{obs_dt}{Data table containing the observations. Pass NULL if the observations are contained in fc_dt.}

\item{obs_cn}{column name of the observations (either in obs_dt, or in fc_dt if obs_dt = NULL).}

\item{by_cns}{column names of grouping variables, all of which need to be columns in fc_dt.
Default is to group by all instances of month, season, lon, lat, system and lead_time that are columns in fc_dt.}

\item{along_cns}{column name(s) for the variable(s) along which is averaged, typically just 'year'.}

\item{member_cn}{Name of the column identifying the ensemble member.}

\item{check_dimension}{Logical. If True, a simple test whether the dimensions match up is conducted:
Namely, getting the mean grouped by by_cns AND along_cns should correspond to getting the ensemble mean. This is tested.}
}
\description{
Taking CRPSs of ensemble forecasts stored in long data tables:
}
