% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confer_validation.R
\name{crps_sample_na}
\alias{crps_sample_na}
\title{Auxiliary function for computing CRPS on ensemble forecasts
Essentially wraps scoringRules::crps_sample, but can deal with missing values.
That's important when your ensemble switches size between hind- and forecasts, and
when you're computing crps for several systems at once and are too lazy to peel them apart.}
\usage{
crps_sample_na(obs, pred, ens_size_correction = FALSE)
}
\arguments{
\item{obs}{vector of observations.}

\item{pred}{either vector of predictions when a single observation is provided, or a matrix with nrow = length(obs)
where the different columns are the different predictions.}

\item{ens_size_correction}{logical. If TRUE, the correction for ensemble size proposed in Ferro et al. 2008: 'On the effect of ensemble size on the discrete and continuous
ranked probability scores' is employed. This makes calculations slower, but should be done when you compare ensembles with different sizes!}
}
\description{
Auxiliary function for computing CRPS on ensemble forecasts
Essentially wraps scoringRules::crps_sample, but can deal with missing values.
That's important when your ensemble switches size between hind- and forecasts, and
when you're computing crps for several systems at once and are too lazy to peel them apart.
}
