% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confer_validation.R
\name{CRPSS_ens_fc}
\alias{CRPSS_ens_fc}
\title{Function for taking CRPS skill scores of ensemble forecasts stored in long data tables:}
\usage{
CRPSS_ens_fc(
  fc_dt,
  fc_cn,
  obs_cn = "obs",
  by_cns = intersect(c("month", "season", "lon", "lat", "system", "lead_time"),
    names(fc_dt)),
  along_cns = c("year"),
  ...
)
}
\arguments{
\item{fc_dt}{Data table containing the predictions.}

\item{fc_cn}{column name of the prediction.}

\item{obs_cn}{column name of the observations (either in obs_dt, or in fc_dt if obs_dt = NULL).}

\item{by_cns}{column names of grouping variables, all of which need to be columns in fc_dt. A separate CRPS is computed for each value of the grouping variables.
Default is to group by all instances of month, season, lon, lat, system and lead_time that are columns in fc_dt.}

\item{along_cns}{column name(s) for the variable(s) along which is averaged. Needs to contain 'year' since the reference climatology forecast is leave-one-year-out.}

\item{...}{passed on to CRPS_ens_fc}

\item{obs_dt}{Data table containing the observations. Pass NULL if the observation are contained in fc_dt.}
}
\description{
Warning: The skill score needs a climatological forecast as reference. This is so far always based on the leave-one-year-out climatology.
}
