% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirps.R
\name{delete_redundant_files}
\alias{delete_redundant_files}
\title{Auxiliary function cleaning out the directories, called at the end of the CHIRPS download.}
\usage{
delete_redundant_files(dir)
}
\arguments{
\item{dir}{the directory of the high dimensional CHIRPS data.}
}
\description{
Auxiliary function cleaning out the directories, called at the end of the CHIRPS download.
}
