% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{dimvars}
\alias{dimvars}
\title{Return the default column names that are considered to be coordinates rather than values:}
\usage{
dimvars(dt = NULL)
}
\description{
Return the default column names that are considered to be coordinates rather than values:
}
