% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{dimvars}
\alias{dimvars}
\title{Get dimension variables}
\usage{
dimvars(dt = NULL)
}
\arguments{
\item{dt}{Optional data table. If a data table is provided only the dimvars of the data table are returned.}
}
\value{
A vector of characters with the column names considered dimvars.
}
\description{
The function returns all names currently considered dimension variables.
Following the logic of netcdfs, data tables usually have columns specifying coordinates
(or dimvars) and other columns containing data for these dimvars. Dimension variables can be spatial
or temporal coordinates, or the lead time of a forecast or the member in an ensemble forecast, etc...
}
\examples{
dimvars()

}
