% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf_to_dt.R
\name{dt_to_netcdf}
\alias{dt_to_netcdf}
\title{Write a netcdf from a long data tables.}
\usage{
dt_to_netcdf(
  dt,
  nc_out,
  vars = NULL,
  units = NULL,
  dim_vars = dim_vars(dt),
  dim_var_units = NULL,
  check = TRUE,
  description = NULL
)
}
\arguments{
\item{dt}{a data.table}

\item{nc_out}{File name (including path) of the netcdf to write.}

\item{vars}{names of columns in dt containing variables. If this is NULL, the function guesses and asks for confirmation.}

\item{units}{character vector containing the units for vars (in the same order). If this is NULL (default), the user is prompted for input.}

\item{dim_vars}{names of columns in dt containing dimension variables. If this is NULL, the function guesses and asks for confirmation.}

\item{dim_var_units}{character vector containing the units for dim_vars (in the same order). If this is NULL (default), the user is prompted for input (except for lon/lat).}

\item{check}{Only used when a file with the given name already exists. Default is to prompt user for input. This can be avoided (e.g. if you automatically want to overwrite a lot of files) by setting check = 'y'.}

\item{description}{For adding a global attribute 'Description' as a string.}
}
\value{
none.
}
\description{
This function writes a netcdf from a long data table, the usual data format in SeaVal.
The function currently does not support writing netcdfs with multiple variables that have different sets of dimension variables!
}
\author{
Claudio
}
