% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecmwf_monthly}
\alias{ecmwf_monthly}
\title{Monthly mean precipitation forecast example dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 37224 rows and 9 columns.
}
\source{
\url{https://cds.climate.copernicus.eu}
}
\usage{
data(ecmwf_monthly)
}
\description{
This is a small example dataset containing hindcasts of monthly mean precipitation for illustration purposes.
The forecasts are contained for the entire GHA-region, for November and December 2018-2020.
The forecasts are issued by the ECMWF SEAS 5 model and initialized in August.
The unit of precipitation is mm/day. Only the first 3 ensemble members are
provided. The dataset also contains tercile probability forecasts, which are
derived from the full 51 member ensemble. The probability for a tercile for a
given year, month and location is always computed as the fraction of ensemble
members falling into that tercile, computed from all ensemble predictions for
the month and location under consideration.
This dataset was generated using Copernicus Climate Change Service information (2020).
}
\keyword{datasets}
