% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{get_terciles}
\alias{get_terciles}
\title{get terciles from a data table}
\usage{
get_terciles(...)
}
\arguments{
\item{...}{passed on to get_quantiles.}
}
\value{
Nothing if save_file is provided. Otherwise the list described above
}
\description{
This function wraps \code{get_quantiles} with the fixed quantiles 0.33 and 0.67.
}
