% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICPAC_temp.R
\name{ggplot_dt_shf}
\alias{ggplot_dt_shf}
\title{Plotting function with different GHA-borders}
\usage{
ggplot_dt_shf(..., expand.x = c(-0.5, 0.5), expand.y = c(-0.5, 2))
}
\arguments{
\item{...}{passed to \link{\code ggplot_dt}}
}
\description{
This loads a (pre-processed) shapefile from ICPAC containing the GHA-countries.
}
