% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot_dt_shf}
\alias{ggplot_dt_shf}
\title{Synonym for \code{\link{ggplot_dt_gha_map}}.}
\usage{
ggplot_dt_shf(..., expand.x = c(-0.5, 0.5), expand.y = c(-0.5, 2))
}
\arguments{
\item{..., expand.x, expand.y}{passed to \code{\link{ggplot_dt}}}
}
\description{
Synonym for \code{\link{ggplot_dt_gha_map}}.
}
