% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confer_validation.R
\name{MBSS_dt}
\alias{MBSS_dt}
\title{Compute the Multicategory Brier Skill score}
\usage{
MBSS_dt(
  fc_dt,
  fc_cols = c("below", "normal", "above"),
  obs_col = "tercile_cat",
  by_cols = intersect(c("month", "season", "lon", "lat", "system", "lead_time"),
    names(fc_dt)),
  along_cols = c("year")
)
}
\arguments{
\item{fc_dt}{Data table containing the predictions.}

\item{fc_cols}{column names of the prediction.}

\item{obs_col}{column name of the observations (either in obs_dt, or in fc_dt if obs_dt = NULL). The observation column needs to
contain -1 if it falls into the first category (corresponding to fc_cols[1]), 0 for the second and 1 for the third category.}

\item{by_cols}{column names of grouping variables, all of which need to be columns in fc_dt.
Default is to group by all instances of month, season, lon, lat, system and lead_time that are columns in fc_dt.}

\item{along_cns}{column name(s) for the variable(s) along which is averaged, typically just 'year'.}
}
\description{
This score is suitable for tercile category forecasts.
}
