% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{obs_dimvars}
\alias{obs_dimvars}
\title{Auxiliary function returning observation dimvars.}
\usage{
obs_dimvars(dt = NULL)
}
\arguments{
\item{dt}{optional. You can provide a data table, then the function returns the names of coordinate columns in this data table.}
}
\value{
Character vector with column names.
}
\description{
Observation dimvars are column names in a data table that resemble coordinates for which only one observation may exist.
}
\examples{
obs_dimvars

}
