% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restrict_to_confer_region}
\alias{restrict_to_confer_region}
\title{restricts data to CONFER region}
\usage{
restrict_to_confer_region(dt, ...)

restrict_to_confer_region(dt, ...)
}
\arguments{
\item{dt}{the data table.}

\item{...}{passed on to \code{\link{restrict_to_country}}}
}
\description{
Wraps \code{\link{restrict_to_country}}, and restricts to the GHA-region usually considered in CONFER, see \code{\link{EA_country_names}}.

Wraps \code{\link{restrict_to_country}}, and restricts to the GHA-region usually considered in CONFER, see \code{\link{EA_country_names}}.
}
