% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restrict_to_country}
\alias{restrict_to_country}
\title{restricts data to a specified country
Only the following 10 countries can be used in this function: Sudan, South Sudan, Somalia, Eritrea, Ethiopia, Somalia, Kenya, Tansania, Uganda, Rwanda, Burundi.
This function assumes the lon/lat grid to be a regular half degree or full degree grid.}
\usage{
restrict_to_country(dt, ct, rectangle = FALSE, tol = 0.5)
}
\arguments{
\item{dt}{the data table.}

\item{ct}{name of the country, or vector containing multiple country names}

\item{rectangle}{logical. If FALSE (default), the data is restricted to the gridcells for which the centerpoint lies within the selected country (e.g. for computing mean scores for a country).
If TRUE, the data is kept for a rectangle containing the entire country, therefore also containing gridpoints outside the country. This is the preferred option for plotting data
for a specific country.}

\item{tol}{Only used when \code{rectangle == TRUE}. A tolerance value for widening the plotting window, making things look a bit nicer.}
}
\description{
restricts data to a specified country
Only the following 10 countries can be used in this function: Sudan, South Sudan, Somalia, Eritrea, Ethiopia, Somalia, Kenya, Tansania, Uganda, Rwanda, Burundi.
This function assumes the lon/lat grid to be a regular half degree or full degree grid.
}
