% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{run_dimension_check_terc_forecast}
\alias{run_dimension_check_terc_forecast}
\title{Auxiliary Function}
\usage{
run_dimension_check_terc_forecast()
}
\description{
called inside functions that calculate scores
for ensemble forecasts. Checks whether the provided data table has the right format.
}
