% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{space_dimvars}
\alias{space_dimvars}
\title{Auxiliary function}
\usage{
space_dimvars(dt = NULL)
}
\arguments{
\item{dt}{optional. You can provide a data table, then the function returns the names of spatial coordinate columns in this data table.}
}
\value{
Character vector with column names.
}
\description{
returns all column names indicating a spatial coordinate.
}
\examples{
space_dimvars()
}
