% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{tc_cols}
\alias{tc_cols}
\title{Tercile column names}
\usage{
tc_cols(dt = NULL)
}
\arguments{
\item{dt}{optional data table. If provided, the function guesses which column contains the observations. Else it returns all recognized column names.}
}
\value{
Character vector with column names.
}
\description{
which column names are interpreted as observed tercile categories
}
\examples{
tc_cols()

}
