% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upscale_nested_griddings}
\alias{upscale_nested_griddings}
\title{upscaling data from a finer to a coarser grid (by averaging) assumes the finer grid to be nested within the larger (only averages )
Does not account for error by projection (lon/lat squares are treated as actual squares), so only use this near the equator.}
\usage{
upscale_nested_griddings(
  dt,
  uscol,
  coarse_grid,
  bycols = intersect(c("month", "year"), colnames(dt)),
  type = "within"
)
}
\description{
upscaling data from a finer to a coarser grid (by averaging) assumes the finer grid to be nested within the larger (only averages )
Does not account for error by projection (lon/lat squares are treated as actual squares), so only use this near the equator.
}
